Entrega backend3

+++ src/services/user.services.js
@@ -4,11 +4,12 @@ import { generateUsersMock } from "../mocks/user.mock.js";
 
 export class UserServices {
   constructor() {
-    // this.userRepository = new UserRepository(new Users());
     this.userDao = new Users();
   }
+
   async getAll() {
     const users = await this.userDao.get();
+    if (!users) throw customError.notFoundError(`No user found`);
     return users;
   }
   async getById(id) {
@@ -18,25 +19,39 @@ export class UserServices {
   }
   async create(data) {
     const user = await this.userDao.save(data);
+    if (!user) throw customError.badRequestError("Failed to create user");
     return user;
   }
 
-  async createMany(data) {
+  async crateMany(data) {
     const users = await this.userDao.saveMany(data);
+
     return users;
   }
 
   async update(id, data) {
     const user = await this.userDao.update(id, data);
+    if (!user) throw customError.notFoundError(`User id ${id} not found`);
     return user;
   }
   async remove(id) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw customError.badRequestError(`Invalid ObjectId: ${id}`);
+    }
+    const user = await this.userDao.getBy(id);
+    if (!user) {
+      throw customError.notFoundError(`User id ${id} not found`);
+    }
     await this.userDao.delete(id);
-    return "ok";
+    return "User Deleted";
   }
+
   async createMocks() {
     const users = generateUsersMock(10);
     const usersDb = await this.userDao.saveMany(users);
+    if (!usersDb)
+      throw customError.badRequestError("Failed to create user mocks");
+
     return usersDb;
   }
 }
